# js: Optimize zxyToTileId performance