This PR improves the performance of `zxyToTileId` by roughly 13-14% via the manual pre-computation and inclusion of the possible `acc` values for zoom. I broke out #132 in a separate PR so that the diff should be easier to digest for this if that's merged first. 

### Benchmarks

Below code requires [benchmark.js](https://www.npmjs.com/package/benchmark).

<details>
 
  <summary>Test Code</summary>

```javascript

const Benchmark = require("benchmark");
var suite = new Benchmark.Suite();

function rotate(n, xy, rx, ry) {
  if (ry == 0) {
    if (rx == 1) {
      xy[0] = n - 1 - xy[0];
      xy[1] = n - 1 - xy[1];
    }
    const t = xy[0];
    xy[0] = xy[1];
    xy[1] = t;
  }
}

const tzValues = [
  0, 1, 5, 21, 85, 341, 1365, 5461, 21845, 87381, 349525, 1398101, 5592405,
  22369621, 89478485, 357913941, 1431655765, 5726623061, 22906492245,
  91625968981, 366503875925, 1466015503701, 5864062014805, 23456248059221,
  93824992236885, 375299968947541, 1501199875790165,
];

function zxyToTileIdNew(z, x, y) {
  if (z > 26) {
    throw Error("Tile zoom level exceeds max safe number limit (26)");
  }
  if (x > Math.pow(2, z) - 1 || y > Math.pow(2, z) - 1) {
    throw Error("tile x/y outside zoom level bounds");
  }

  const acc = tzValues[z];
  const n = Math.pow(2, z);
  let rx = 0;
  let ry = 0;
  let d = 0;
  const xy = [x, y];
  let s = n / 2;
  while (s > 0) {
    rx = (xy[0] & s) > 0 ? 1 : 0;
    ry = (xy[1] & s) > 0 ? 1 : 0;
    d += s * s * ((3 * rx) ^ ry);
    rotate(s, xy, rx, ry);
    s = s / 2;
  }
  return acc + d;
}

function zxyToTileIdOld(z, x, y) {
  if (z > 26) {
    throw Error("Tile zoom level exceeds max safe number limit (26)");
  }
  if (x > Math.pow(2, z) - 1 || y > Math.pow(2, z) - 1) {
    throw Error("tile x/y outside zoom level bounds");
  }

  let acc = 0;
  let tz = 0;
  while (tz < z) {
    acc += Math.pow(2, tz) * Math.pow(2, tz);
    tz++;
  }
  const n = Math.pow(2, z);
  let rx = 0;
  let ry = 0;
  let d = 0;
  const xy = [x, y];
  let s = n / 2;
  while (s > 0) {
    rx = (xy[0] & s) > 0 ? 1 : 0;
    ry = (xy[1] & s) > 0 ? 1 : 0;
    d += s * s * ((3 * rx) ^ ry);
    rotate(s, xy, rx, ry);
    s = s / 2;
  }
  return acc + d;
}

// add tests
suite
  .add("old", function () {
    zxyToTileIdOld(8, 0, 0);
  })
  .add("new", function () {
    zxyToTileIdNew(8, 0, 0);
  })
  // add listeners
  .on("cycle", function (event) {
    console.log(String(event.target));
  })
  .on("complete", function () {
    console.log("Fastest is " + this.filter("fastest").map("name"));
  })
  // run async
  .run({ async: true });

```
</details>

#### Results
```md
old x 510,480 ops/sec ±0.31% (99 runs sampled)
new x 580,460 ops/sec ±0.30% (97 runs sampled)
Fastest is new
```

Shoutout to @sheetjsdev for helping me out with the bitwise operations here as this is new territory for me.