@bdon 2023-03-13
Nice optimization! Thanks!